moved iterators and spliterators from Img class to Iterators class; refactored a lot of the Img methods to a new base interface BaseImg (contains all the iterator/foreach/stream code)
