+++ ImagingKit_Core/src/main/java/hageldave/imagingkit/core/Img.java
@@ -23,7 +23,6 @@
 package hageldave.imagingkit.core;
 
 import java.awt.Dimension;
-import java.awt.Graphics2D;
 import java.awt.image.BufferedImage;
 import java.awt.image.DataBuffer;
 import java.awt.image.DataBufferInt;
@@ -31,16 +30,8 @@ import java.awt.image.DirectColorModel;
 import java.awt.image.Raster;
 import java.awt.image.WritableRaster;
 import java.util.Arrays;
-import java.util.Iterator;
 import java.util.Spliterator;
 import java.util.function.Consumer;
-import java.util.function.Supplier;
-import java.util.stream.Stream;
-import java.util.stream.StreamSupport;
-
-import hageldave.imagingkit.core.PixelConvertingSpliterator.PixelConverter;
-import hageldave.imagingkit.core.util.BufferedImageFactory;
-import hageldave.imagingkit.core.util.ParallelForEachExecutor;
 
 /**
  * Image class with data stored in an int array.
@@ -95,7 +86,7 @@ import hageldave.imagingkit.core.util.ParallelForEachExecutor;
  * @author hageldave
  * @since 1.0
  */
-public class Img implements Iterable<Pixel> {
+public class Img implements ImgBase<Pixel> {
 
 	/** boundary mode that will return 0 for out of bounds positions.
 	 * @see #getValue(int, int, int)
@@ -509,27 +500,7 @@ public class Img implements Iterable<Pixel> {
 		return new Img(getDimension(), Arrays.copyOf(getData(), getData().length));
 	}
 
-	/**
-	 * @return a BufferedImage of type INT_ARGB with this Img's data copied to it.
-	 * @see #toBufferedImage(BufferedImage)
-	 * @see #getRemoteBufferedImage()
-	 * @since 1.0
-	 */
-	public BufferedImage toBufferedImage(){
-		BufferedImage bimg = BufferedImageFactory.getINT_ARGB(getDimension());
-		return toBufferedImage(bimg);
-	}
-
-	/**
-	 * Copies this Img's data to the specified {@link BufferedImage}.
-	 * @param bimg the BufferedImage
-	 * @return specified BufferedImage
-	 * @throws ArrayIndexOutOfBoundsException if the provided BufferedImage
-	 * has less values than this Img.
-	 * @see #toBufferedImage()
-	 * @see #getRemoteBufferedImage()
-	 * @since 1.0
-	 */
+	@Override
 	public BufferedImage toBufferedImage(BufferedImage bimg){
 		bimg.setRGB(0, 0, getWidth(), getHeight(), getData(), 0, getWidth());
 		return bimg;
@@ -545,6 +516,7 @@ public class Img implements Iterable<Pixel> {
 	 * @see #toBufferedImage()
 	 * @since 1.0
 	 */
+	@Override
 	public BufferedImage getRemoteBufferedImage(){
 		DirectColorModel cm = new DirectColorModel(32,
 				0x00ff0000,       // Red
@@ -556,6 +528,11 @@ public class Img implements Iterable<Pixel> {
 		WritableRaster raster = Raster.createPackedRaster(buffer, getWidth(), getHeight(), getWidth(), cm.getMasks(), null);
 		BufferedImage bimg = new BufferedImage(cm, raster, false, null);
 		return bimg;
+	}
+
+	@Override
+	public boolean supportsRemoteBufferedImage() {
+		return true;
 	}
 
 	/**
@@ -590,96 +567,6 @@ public class Img implements Iterable<Pixel> {
 	}
 
 
-	@Override
-	public Iterator<Pixel> iterator() {
-		return new ImgIterator<Pixel>(numValues(), getPixel()) ;
-	}
-
-	/**
-	 * Returns an iterator for the specified area of the image.
-	 * @param xStart left boundary of the area (inclusive)
-	 * @param yStart upper boundary of the area (inclusive)
-	 * @param width of the area
-	 * @param height of the area
-	 * @return iterator for iterating over the pixels in the specified area.
-	 * @throws IllegalArgumentException if provided area is not within this
-	 * Img's bounds.
-	 * @since 1.1
-	 */
-	public Iterator<Pixel> iterator(final int xStart, final int yStart, final int width, final int height) {
-		if(		width <= 0 || height <= 0 ||
-				xStart < 0 || yStart < 0 ||
-				xStart+width > getWidth() || yStart+height > getHeight() )
-		{
-			throw new IllegalArgumentException(String.format(
-							"provided area [%d,%d][%d,%d] is not within bounds of the image [%d,%d]",
-							xStart,yStart,width,height, getWidth(), getHeight()));
-		}
-		return new ImgAreaIterator<Pixel>(xStart, yStart, width, height, getPixel());
-	}
-
-	@Override
-	public Spliterator<Pixel> spliterator() {
-		return new ImgSpliterator<Pixel>(0, numValues()-1, spliteratorMinimumSplitSize,this::getPixel);
-	}
-
-	/**
-	 * Creates a {@link Spliterator} that guarantees that each split will
-	 * at least cover an entire row of the Img. It also guarantes that each
-	 * row will be iterated starting at the least index of that row
-	 * (e.g.starts at index 0 then continues with index 1, then 2, until
-	 * the end of the row, then continuing with the next row).
-	 * This Spliterator iterates in row-major order.
-	 * @return Spliterator that splits at beginning of rows.
-	 * @see #colSpliterator()
-	 * @see #spliterator()
-	 * @see #stream(Spliterator, boolean)
-	 * @since 1.3
-	 */
-	public Spliterator<Pixel> rowSpliterator() {
-		return new RowSpliterator<Pixel>(0, getWidth(), 0, getHeight()-1, this::getPixel);
-	}
-
-	/**
-	 * Creates a {@link Spliterator} that guarantees that each split will
-	 * at least cover an entire column of the Img. It also guarantes that each
-	 * column will be iterated starting at the least index of that column
-	 * (e.g.starts at index 0 then continues with index 1, then 2, until
-	 * the end of the column, then continuing with the next column).
-	 * This Spliterator iterates in column-major order.
-	 * @return Spliterator that splits at beginning of columns.
-	 * @see #rowSpliterator()
-	 * @see #spliterator()
-	 * @see #stream(Spliterator, boolean)
-	 * @since 1.3
-	 */
-	public Spliterator<Pixel> colSpliterator() {
-		return new ColSpliterator<Pixel>(0, getWidth()-1, 0, getHeight(), this::getPixel);
-	}
-
-	/**
-	 * Creates a {@link Spliterator} over the pixels within the specified area.
-	 * @param xStart left boundary of the area (inclusive)
-	 * @param yStart upper boundary of the area (inclusive)
-	 * @param width of the area
-	 * @param height of the area
-	 * @return spliterator for the specified area.
-	 * @throws IllegalArgumentException if provided area is not within this
-	 * Img's bounds.
-	 * @since 1.1
-	 */
-	public Spliterator<Pixel> spliterator(final int xStart, final int yStart, final int width, final int height) {
-		if(		width <= 0 || height <= 0 ||
-				xStart < 0 || yStart < 0 ||
-				xStart+width > getWidth() || yStart+height > getHeight() )
-		{
-			throw new IllegalArgumentException(String.format(
-							"provided area [%d,%d][%d,%d] is not within bounds of the image [%d,%d]",
-							xStart,yStart,width,height, getWidth(), getHeight()));
-		}
-		return new ImgAreaSpliterator<Pixel>(xStart,yStart,width,height, spliteratorMinimumSplitSize, this::getPixel);
-	}
-
 	/**
 	 * Returns the minimum number of elements in a split of a {@link Spliterator}
 	 * of this Img. Spliterators will only split if they contain more elements than
@@ -687,6 +574,7 @@ public class Img implements Iterable<Pixel> {
 	 * @return minimum number of elements of a Spliterator to allow for splitting.
 	 * @since 1.3
 	 */
+	@Override
 	public int getSpliteratorMinimumSplitSize() {
 		return spliteratorMinimumSplitSize;
 	}
@@ -713,704 +601,6 @@ public class Img implements Iterable<Pixel> {
 					String.format("Minimum split size has to be above zero, specified:%d", size));
 		}
 		this.spliteratorMinimumSplitSize = size;
-	}
-
-
-	/** default implementation of {@link Iterable#forEach(Consumer)} <br>
-	 * only for performance test purposes as it is slower than the
-	 * {@link Img#forEach(Consumer)} implementation
-	 * @since 1.0
-	 */
-	void forEach_defaultimpl(final Consumer<? super Pixel> action) {
-		Iterable.super.forEach(action);
-	}
-
-	/**
-	 * @see #forEachParallel(Consumer action)
-	 * @since 1.0
-	 */
-	@Override
-	public void forEach(final Consumer<? super Pixel> action) {
-		forEach(false, action);
-	}
-	
-	public void forEach(boolean parallel, final Consumer<? super Pixel> action) {
-		if(parallel){
-			ParallelForEachExecutor<Pixel> exec = new ParallelForEachExecutor<>(null, spliterator(), action);
-			exec.invoke();
-		} else {
-			Pixel p = getPixel();
-			for(int i = 0; i < numValues(); p.setIndex(++i)){
-				action.accept(p);
-			}
-		}
-	}
-	
-	/**
-	 * Applies the specified action to every pixel in the specified area of this image.
-	 * @param xStart left boundary of the area (inclusive)
-	 * @param yStart upper boundary of the area (inclusive)
-	 * @param width of the area
-	 * @param height of the area
-	 * @param action to be performed on each pixel
-	 * @throws IllegalArgumentException if provided area is not within this
-	 * Img's bounds.
-	 * @see #forEachParallel(int x, int y, int w, int h, Consumer action)
-	 * @since 1.1
-	 */
-	public void forEach(final int xStart, final int yStart, final int width, final int height, final Consumer<? super Pixel> action) {
-		forEach(false, xStart, yStart, width, height, action);
-	}
-
-	public void forEach(boolean parallel, final int xStart, final int yStart, final int width, final int height, final Consumer<? super Pixel> action) {
-		if(parallel){
-			ParallelForEachExecutor<Pixel> exec = new ParallelForEachExecutor<>(null, spliterator(xStart, yStart, width, height), action);
-			exec.invoke();
-		} else {
-			Pixel p = getPixel();
-			int yEnd = yStart+height;
-			int xEnd = xStart+width;
-			for(int y = yStart; y < yEnd; y++){
-				for(int x = xStart; x < xEnd; x++){
-					p.setPosition(x, y);
-					action.accept(p);
-				}
-			}
-		}
-	}
-
-	
-	public <T> void forEach(final PixelConverter<T> converter, boolean parallel, final Consumer<? super T> action) {
-		if(parallel){
-			Spliterator<T> spliterator = new PixelConvertingSpliterator<T>(
-					spliterator(), 
-					converter);
-	 		ParallelForEachExecutor<T> exec = new ParallelForEachExecutor<>(null, spliterator, action);
-			exec.invoke();
-		} else {
-			Pixel px = getPixel();
-			T element = converter.allocateElement();
-			for(int i = 0; i < numValues(); px.setIndex(++i)){
-				converter.convertPixelToElement(px, element);
-				action.accept(element);
-				converter.convertElementToPixel(element, px);
-			}
-		}
-	}
-
-	
-	public <T> void forEach(final PixelConverter<T> converter, boolean parallel, final int xStart, final int yStart, final int width, final int height, final Consumer<? super T> action) {
-		if(parallel){
-			Spliterator<T> spliterator = new PixelConvertingSpliterator<T>(
-					spliterator(xStart, yStart, width, height), 
-					converter);
-			ParallelForEachExecutor<T> exec = new ParallelForEachExecutor<>(null, spliterator, action);
-			exec.invoke();
-		} else {
-			Pixel p = getPixel();
-			T element = converter.allocateElement();
-			int yEnd = yStart+height;
-			int xEnd = xStart+width;
-			for(int y = yStart; y < yEnd; y++){
-				for(int x = xStart; x < xEnd; x++){
-					p.setPosition(x, y);
-					converter.convertPixelToElement(p, element);
-					action.accept(element);
-					converter.convertElementToPixel(element, p);
-				}
-			}
-		}
-	}
-
-	/**
-	 * Returns a {@code Stream<Pixel>} for the specified {@code Spliterator<Pixel>}.
-	 * This is just a wrapper method arround {@link StreamSupport#stream(Spliterator, boolean)}
-	 * mainly used as syntactic sugar for the use with the non default Spliterators
-	 * ({@link #rowSpliterator()} and {@link #colSpliterator()}. When the default spliterator
-	 * of the Img is sufficient the non-static {@link #stream()} can be used.
-	 * <p>
-	 * For example (horizontal edge detection in parallel using forward difference):
-	 * <pre>
-	 * {@code
-	 * Img myImg = ...;
-	 * Img.stream(myImg.rowSpliterator(), true).forEach( px -> {
-	 *     int next = px.getImg().getValue(px.getX()+1, px.getY(), Img.boundary_mode_repeat_edge);
-	 *     int forwardDiff = Math.abs( Pixel.getLuminance(next) - px.getLuminance() );
-	 *     px.setRGB(forwardDiff, forwardDiff, forwardDiff);
-	 * });
-	 * }
-	 * </pre>
-	 * @param spliterator Spliterator of Img to be streamed
-	 * @param parallel whether parallel or sequential stream is returned
-	 * @return a new sequential or parallel pixel stream.
-	 *
-	 * @see #stream()
-	 * @see #parallelStream()
-	 * @since 1.3
-	 */
-	public static Stream<Pixel> stream(Spliterator<Pixel> spliterator, boolean parallel){
-		return StreamSupport.stream(spliterator, parallel);
-	}
-
-	/**
-	 * Returns a Pixel {@link Stream} of this Img.
-	 * This Img's {@link #spliterator()} is used to create the Stream.
-	 * @return Pixel Stream of this Img.
-	 * @see #parallelStream()
-	 * @see #stream(int x, int y, int w, int h)
-	 * @since 1.2
-	 */
-	public Stream<Pixel> stream() {
-		return stream(false);
-	}
-	
-	public Stream<Pixel> stream(boolean parallel) {
-		return Img.stream(spliterator(), parallel);
-	}
-	
-	/**
-	 * Returns a Pixel {@link Stream} for the specified area of this Img.<br>
-	 * This Img's {@link #spliterator(int,int,int,int)} is used to create
-	 * the Stream.
-	 * @param xStart left boundary of the area (inclusive)
-	 * @param yStart upper boundary of the area (inclusive)
-	 * @param width of the area
-	 * @param height of the area
-	 * @return Pixel Stream for specified area.
-	 * @throws IllegalArgumentException if provided area is not within this
-	 * Img's bounds.
-	 * @see #parallelStream(int x, int y, int w, int h)
-	 * @see #stream()
-	 * @since 1.2
-	 */
-	public Stream<Pixel> stream(final int xStart, final int yStart, final int width, final int height){
-		return stream(false, xStart, yStart, width, height);
-	}
-
-	public Stream<Pixel> stream(boolean parallel, final int xStart, final int yStart, final int width, final int height){
-		return StreamSupport.stream(spliterator(xStart, yStart, width, height), parallel);
-	}
-
-	public <T> Stream<T> stream(PixelConverter<T> converter, boolean parallel) {
-		Spliterator<T> spliterator = new PixelConvertingSpliterator<T>(spliterator(), converter);
-		return StreamSupport.stream(spliterator, parallel);
-	}
-
-	public <T> Stream<T> stream(final PixelConverter<T> converter, boolean parallel, final int xStart, final int yStart, final int width, final int height){
-		Spliterator<T> spliterator = new PixelConvertingSpliterator<>(
-				spliterator(xStart, yStart, width, height), 
-				converter);
-		return StreamSupport.stream(spliterator, parallel);
-	}
-
-	/**
-	 * Creates a {@link Graphics2D}, which can be used to draw into this Img.
-	 * @return Graphics2D object to draw into this image.
-	 * @see #paint(Consumer)
-	 * @since 1.3
-	 */
-	public Graphics2D createGraphics(){
-		return getRemoteBufferedImage().createGraphics();
-	}
-
-	/**
-	 * Uses the specified paintInstructions to draw into this Img.
-	 * This method will pass a {@link Graphics2D} object of this Img to the
-	 * specified {@link Consumer}. The {@link Consumer#accept(Object)} method
-	 * can then draw into this Image. When the accept method return, the
-	 * Graphics2D object is disposed.
-	 * <p>
-	 * For example (using lambda expression for Consumers accept method):
-	 * <pre>
-	 * {@code
-	 * Img img = new Img(100, 100);
-	 * img.paint( g2d -> { g2d.drawLine(0, 0, 100, 100); } );
-	 * }
-	 * </pre>
-	 * @param paintInstructions to be executed on a this Graphics2D object
-	 * of this Img.
-	 * @see #createGraphics()
-	 * @since 1.3
-	 */
-	public void paint(Consumer<Graphics2D> paintInstructions){
-		Graphics2D g2d = createGraphics();
-		paintInstructions.accept(g2d);
-		g2d.dispose();
-	}
-
-	
-	public static class ImgIterator<T extends PixelBase> implements Iterator<T> {
-		final T px;
-		final int numValues;
-		int index;
-		
-		public ImgIterator(final int numValues, T px) {
-			this.px = px;
-			this.numValues = numValues;
-			this.index = -1;
-		}
-
-		@Override
-		public T next() {
-			px.setIndex(++index);
-			return px;
-		}
-
-		@Override
-		public boolean hasNext() {
-			return index+1 < numValues;
-		}
-
-		@Override
-		public void forEachRemaining(Consumer<? super T> action) {
-			px.setIndex(++index);
-			for(/*index*/; index < numValues; px.setIndex(++index)){
-				action.accept(px);
-			}
-		}
-	}
-	
-	
-	public static class ImgAreaIterator<T extends PixelBase> implements Iterator<T> {
-		private final T px;
-		private final int xStart; 
-		private final int yStart; 
-		private final int width; 
-		private final int height;
-		int x;
-		int y;
-		
-		public ImgAreaIterator(final int xStart, final int yStart, final int width, final int height, T px) {
-			this.px = px;
-			this.xStart = xStart;
-			this.yStart = yStart;
-			this.width = width;
-			this.height = height;
-			this.x=0;
-			this.y=0;
-		}
-		
-		
-		@Override
-		public T next() {
-			px.setPosition(x+xStart, y+yStart);
-			x++;
-			if(x >= width){
-				x=0;
-				y++;
-			}
-			return px;
-		}
-
-		@Override
-		public boolean hasNext() {
-			return y < height;
-		}
-
-		@Override
-		public void forEachRemaining(Consumer<? super T> action) {
-			int xEnd = xStart+width;
-			int yEnd = yStart+height;
-			int x = this.x+xStart;
-			for(int y=this.y+yStart; y < yEnd; y++){
-				for(; x < xEnd; x++){
-					px.setPosition(x, y);
-					action.accept(px);
-				}
-				x = xStart;
-			}
-		}
-	}
-	
-
-	/**
-	 * Spliterator class for Img bound to a specific area
-	 * @author hageldave
-	 * @since 1.1
-	 */
-	public static final class ImgAreaSpliterator<T extends PixelBase> implements Spliterator<T> {
-
-		private final Supplier<T> pixelSupplier;
-		private final T px;
-		/* start x coord and end x coord of a row */
-		private final int startX, endXexcl;
-		/* current coords of this spliterator */
-		private int x,y;
-		/* final coords of this spliterator */
-		private int finalXexcl, finalYincl;
-
-		private final int minimumSplitSize;
-
-		/**
-		 * Constructs a new ImgAreaSpliterator for the specified area
-		 * @param xStart left boundary of the area (inclusive)
-		 * @param yStart upper boundary of the area (inclusive)
-		 * @param width of the area
-		 * @param height of the area
-		 * @param minSplitSize the minimum number of elements in a split
-		 * @since 1.1
-		 */
-		public ImgAreaSpliterator(
-				int xStart, 
-				int yStart, 
-				int width, 
-				int height, 
-				int minSplitSize, 
-				Supplier<T> pixelSupplier
-		){
-			this(xStart, xStart+width, xStart, yStart, xStart+width, yStart+height-1, minSplitSize, pixelSupplier);
-		}
-
-		private ImgAreaSpliterator(
-				int xStart, 
-				int endXexcl, 
-				int x, 
-				int y, 
-				int finalXexcl, 
-				int finalYincl, 
-				int minSplitSize, 
-				Supplier<T> pixelSupplier
-		){
-			this.startX = xStart;
-			this.endXexcl = endXexcl;
-			this.x = x;
-			this.y = y;
-			this.finalXexcl = finalXexcl;
-			this.finalYincl = finalYincl;
-			this.pixelSupplier = pixelSupplier;
-			this.px = pixelSupplier.get();
-			this.px.setPosition(x, y);
-			this.minimumSplitSize = minSplitSize;
-		}
-
-
-		@Override
-		public boolean tryAdvance(final Consumer<? super T> action) {
-			if(y > finalYincl || (y == finalYincl && x >= finalXexcl)){
-				return false;
-			} else {
-				action.accept(px);
-				if(x+1 >= endXexcl){
-					x = startX;
-					y++;
-				} else {
-					x++;
-				}
-				px.setPosition(x, y);
-				return true;
-			}
-		}
-
-		@Override
-		public void forEachRemaining(final Consumer<? super T> action) {
-			if(this.y == finalYincl){
-				for(int x = this.x; x < finalXexcl; x++){
-					px.setPosition(x, finalYincl);
-					action.accept(px);
-				}
-			} else {
-				// end current row
-				for(int x = this.x; x < endXexcl; x++){
-					px.setPosition(x, this.y);
-					action.accept(px);
-				}
-				// do next rows right before final row
-				for(int y = this.y+1; y < this.finalYincl; y++){
-					for(int x = startX; x < endXexcl; x++ ){
-						px.setPosition(x, y);
-						action.accept(px);
-					}
-				}
-				// do final row
-				for(int x = startX; x < finalXexcl; x++){
-					px.setPosition(x, finalYincl);
-					action.accept(px);
-				}
-			}
-		}
-
-		@Override
-		public Spliterator<T> trySplit() {
-			int width = (this.endXexcl-this.startX);
-			int idx = this.x - this.startX;
-			int finalIdx_excl = (this.finalYincl-this.y)*width + (this.finalXexcl-startX);
-			int midIdx_excl = idx + (finalIdx_excl-idx)/2;
-			if(midIdx_excl > idx+minimumSplitSize){
-//				int midIdx_excl = idx + (finalIdx_excl-idx)/2;
-
-				int newFinalX_excl = startX + (midIdx_excl%width);
-				int newFinalY_incl = this.y + midIdx_excl/width;
-				ImgAreaSpliterator<T> split = new ImgAreaSpliterator<>(
-						startX,         // start of a row
-						endXexcl,       // end of a row
-						newFinalX_excl, // x coord of new spliterator
-						newFinalY_incl, // y coord of new spliterator
-						finalXexcl,     // final x coord of new spliterator
-						finalYincl,    // final y coord of new spliterator
-						minimumSplitSize,
-						pixelSupplier);
-
-				// shorten this spliterator because new one takes care of the rear part
-				this.finalXexcl = newFinalX_excl;
-				this.finalYincl = newFinalY_incl;
-
-				return split;
-			} else {
-				return null;
-			}
-		}
-
-		@Override
-		public long estimateSize() {
-			int idx = this.x - this.startX;
-			int finalIdx_excl = (this.finalYincl-this.y)*(this.endXexcl-this.startX) + (this.finalXexcl-startX);
-			return finalIdx_excl-idx;
-		}
-
-		@Override
-		public int characteristics() {
-			return NONNULL | SIZED | CONCURRENT | SUBSIZED | IMMUTABLE;
-		}
-
-	}
-
-	/**
-	 * Spliterator class for Img
-	 * @author hageldave
-	 * @since 1.0
-	 */
-	public static final class ImgSpliterator<T extends PixelBase> implements Spliterator<T> {
-
-		private final Supplier<T> pixelSupplier;
-		private final T px;
-		private int endIndex;
-		private final int minimumSplitSize;
-
-		/**
-		 * Constructs a new ImgSpliterator for the specified index range
-		 * @param startIndex first index of the range (inclusive)
-		 * @param endIndex last index of the range (inclusive)
-		 * @param minSplitSize minimum split size for this spliterator (minimum number of elements in a split)
-		 * @param pixelSupplier a function that allocates a new Pixel (T)
-		 * that points to the index given by the function argument
-		 * @since 1.0
-		 */
-		public ImgSpliterator(int startIndex, int endIndex, int minSplitSize, Supplier<T> pixelSupplier) {
-			this.pixelSupplier = pixelSupplier;
-			this.px = pixelSupplier.get();
-			this.px.setIndex(startIndex);
-			this.endIndex = endIndex;
-			this.minimumSplitSize = minSplitSize;
-		}
-
-		private void setEndIndex(int endIndex) {
-			this.endIndex = endIndex;
-		}
-
-		@Override
-		public boolean tryAdvance(final Consumer<? super T> action) {
-			if(px.getIndex() <= endIndex){
-				int index = px.getIndex();
-				action.accept(px);
-				px.setIndex(index+1);
-				return true;
-			} else {
-				return false;
-			}
-		}
-
-		@Override
-		public void forEachRemaining(final Consumer<? super T> action) {
-			int idx = px.getIndex();
-			for(;idx <= endIndex; px.setIndex(++idx)){
-				action.accept(px);
-			}
-		}
-
-		@Override
-		public Spliterator<T> trySplit() {
-			int currentIdx = Math.min(px.getIndex(), endIndex);
-			int midIdx = currentIdx + (endIndex-currentIdx)/2;
-			if(midIdx > currentIdx+minimumSplitSize){
-				ImgSpliterator<T> split = new ImgSpliterator<>(midIdx, endIndex, minimumSplitSize, pixelSupplier);
-				setEndIndex(midIdx-1);
-				return split;
-			} else {
-				return null;
-			}
-		}
-
-		@Override
-		public long estimateSize() {
-			int currentIndex = px.getIndex();
-			int lastIndexPlusOne = endIndex+1;
-			return lastIndexPlusOne-currentIndex;
-		}
-
-		@Override
-		public int characteristics() {
-			return NONNULL | SIZED | CONCURRENT | SUBSIZED | IMMUTABLE;
-		}
-
-	}
-
-	/**
-	 * Special Spliterator which guarantees that each split will cover at least
-	 * an entire row of the image.
-	 * @author hageldave
-	 * @since 1.3
-	 */
-	public static final class RowSpliterator<T extends PixelBase> implements Spliterator<T> {
-
-		private final int startX;
-		private final int endXinclusive;
-		private int x;
-		private int y;
-		private int endYinclusive;
-		private final Supplier<T> pixelSupplier;
-		private final T px;
-
-		public RowSpliterator(int startX, int width, int startY, int endYincl, Supplier<T> pixelSupplier) {
-			this.startX = startX;
-			this.x = startX;
-			this.endXinclusive = startX+width-1;
-			this.y = startY;
-			this.endYinclusive = endYincl;
-			this.pixelSupplier = pixelSupplier;
-			this.px = pixelSupplier.get();
-			this.px.setPosition(x,y);
-		}
-
-
-		@Override
-		public boolean tryAdvance(Consumer<? super T> action) {
-			if(x <= endXinclusive){
-				px.setPosition(x, y);
-				x++;
-			} else if(y < endYinclusive) {
-				y++;
-				x=startX;
-				px.setPosition(x, y);
-				x++;
-			} else {
-				return false;
-			}
-			action.accept(px);
-			return true;
-		}
-
-		@Override
-		public void forEachRemaining(Consumer<? super T> action) {
-			int x_ = x;
-			for(int y_ = y; y_ <= endYinclusive; y_++){
-				for(;x_ <= endXinclusive; x_++){
-					px.setPosition(x_, y_);
-					action.accept(px);
-				}
-				x_=startX;
-			}
-		}
-
-		@Override
-		public Spliterator<T> trySplit() {
-			if(this.y < endYinclusive){
-				int newY = y + 1 + (endYinclusive-y)/2;
-				RowSpliterator<T> split = new RowSpliterator<>(startX, endXinclusive-startX+1, newY, endYinclusive, pixelSupplier);
-				this.endYinclusive = newY-1;
-				return split;
-			} else return null;
-		}
-
-		@Override
-		public long estimateSize() {
-			return (endYinclusive-y)*(endXinclusive+1-startX)+endXinclusive+1-x;
-		}
-
-		@Override
-		public int characteristics() {
-			return NONNULL | SIZED | CONCURRENT | SUBSIZED | IMMUTABLE;
-		}
-
-	}
-
-
-	/**
-	 * Special Spliterator which guarantees that each split will cover at least
-	 * an entire column of the image.
-	 * @author hageldave
-	 * @since 1.3
-	 */
-	public static final class ColSpliterator<T extends PixelBase> implements Spliterator<T> {
-
-		private final int startY;
-		private int endXinclusive;
-		private int x;
-		private int y;
-		private final int endYinclusive;
-		private final Supplier<T> pixelSupplier;
-		private final T px;
-
-		public ColSpliterator(int startX, int endXincl, int startY, int height, Supplier<T> pixelSupplier) {
-			this.startY = startY;
-			this.y = startY;
-			this.endYinclusive = startY+height-1;
-			this.x = startX;
-			this.endXinclusive = endXincl;
-			this.pixelSupplier = pixelSupplier;
-			this.px = pixelSupplier.get();
-			this.px.setPosition(x, y);
-		}
-
-
-		@Override
-		public boolean tryAdvance(Consumer<? super T> action) {
-			if(y <= endYinclusive){
-				px.setPosition(x, y);
-				y++;
-			} else if(x < endXinclusive) {
-				x++;
-				y=startY;
-				px.setPosition(x, y);
-				y++;
-			} else {
-				return false;
-			}
-			action.accept(px);
-			return true;
-		}
-
-		@Override
-		public void forEachRemaining(Consumer<? super T> action) {
-			int y_ = y;
-			for(int x_ = x; x_ <= endXinclusive; x_++){
-				for(;y_ <= endYinclusive; y_++){
-					px.setPosition(x_, y_);
-					action.accept(px);
-				}
-				y_=startY;
-			}
-		}
-
-		@Override
-		public Spliterator<T> trySplit() {
-			if(this.x < endXinclusive){
-				int newX = x + 1 + (endXinclusive-x)/2;
-				ColSpliterator<T> split = new ColSpliterator<>(newX, endXinclusive, startY, endYinclusive-startY+1, pixelSupplier);
-				this.endXinclusive = newX-1;
-				return split;
-			} else return null;
-		}
-
-		@Override
-		public long estimateSize() {
-			return (endXinclusive-x)*(endYinclusive+1-startY)+endYinclusive+1-y;
-		}
-
-		@Override
-		public int characteristics() {
-			return NONNULL | SIZED | CONCURRENT | SUBSIZED | IMMUTABLE;
-		}
-
 	}
